feat: App Starter client working with the apis.
feat: All the services running with a main service which make the task running in parallel.

feat: DimAddress Service, Cargonizer Services and XMLLogistics Services with a IGeneric service which contains the contracts between all the classes.

feat: Printer Helper to print the deals and print the best deal.

feat: Wrapper which contains a retry policy from Polly to handle cases when the services are unavailable or it doesn't even started.

feat: appsettings.json in the console application.